[ARI] Implemented more types and generalized the usage of Integer with Numeric

+++ src/plugins/arithmetic/formPair.go
@@ -30,13 +30,13 @@ type EvaluablePair[T any] interface {
 	GetSymbol() PairOperator
 }
 
-type PairForm[T, U Evaluable[int]] struct {
+type PairForm[T, U Evaluable[Numeric]] struct {
 	first  T
 	second U
 	symbol PairOperator
 }
 
-func NewPairForm[T, U Evaluable[int]](first T, second U, symbol PairOperator) *PairForm[T, U] {
+func NewPairForm[T, U Evaluable[Numeric]](first T, second U, symbol PairOperator) *PairForm[T, U] {
 	return &PairForm[T, U]{first, second, symbol}
 }
 
@@ -55,8 +55,8 @@ func (pf *PairForm[T, U]) Copy() Form {
 	return pf.TrueCopy()
 }
 
-func (pf *PairForm[T, U]) getFactorMap() map[string]int {
-	factorMap := make(map[string]int)
+func (pf *PairForm[T, U]) getFactorMap() map[string]Numeric {
+	factorMap := make(map[string]Numeric)
 	firstChildMap := pf.GetFirst().getFactorMap()
 	secondChildMap := pf.GetSecond().getFactorMap()
 
@@ -71,10 +71,10 @@ func (pf *PairForm[T, U]) getFactorMap() map[string]int {
 	return factorMap
 }
 
-func (pf *PairForm[T, U]) TrueCopy() *PairForm[Evaluable[int], Evaluable[int]] {
-	if typedFirst, ok := pf.first.Copy().(Evaluable[int]); ok {
-		if typedSecond, ok := pf.second.Copy().(Evaluable[int]); ok {
-			return NewPairForm[Evaluable[int], Evaluable[int]](typedFirst, typedSecond, pf.symbol)
+func (pf *PairForm[T, U]) TrueCopy() *PairForm[Evaluable[Numeric], Evaluable[Numeric]] {
+	if typedFirst, ok := pf.first.Copy().(Evaluable[Numeric]); ok {
+		if typedSecond, ok := pf.second.Copy().(Evaluable[Numeric]); ok {
+			return NewPairForm[Evaluable[Numeric], Evaluable[Numeric]](typedFirst, typedSecond, pf.symbol)
 		}
 	}
 
@@ -94,39 +94,39 @@ func (pf *PairForm[T, U]) GetSymbol() PairOperator {
 }
 
 type Sum struct {
-	*PairForm[Evaluable[int], Evaluable[int]]
+	*PairForm[Evaluable[Numeric], Evaluable[Numeric]]
 }
 
-func NewSum(first, second Evaluable[int]) *Sum {
+func NewSum(first, second Evaluable[Numeric]) *Sum {
 	return &Sum{NewPairForm(first, second, SumOperator)}
 }
 
-func NewDiff(first, second Evaluable[int]) *Sum {
-	return &Sum{NewPairForm(first, Evaluable[int](NewNeg(second)), SumOperator)}
+func NewDiff(first, second Evaluable[Numeric]) *Sum {
+	return &Sum{NewPairForm(first, Evaluable[Numeric](NewNeg(second)), SumOperator)}
 }
 
 func (s *Sum) Copy() Form {
 	return &Sum{s.PairForm.TrueCopy()}
 }
 
-var sum = func(first, second int) int {
+var sum = func(first, second Numeric) Numeric {
 	return first + second
 }
 
-var diff = func(first, second int) int {
+var diff = func(first, second Numeric) Numeric {
 	return first - second
 }
 
-func (s *Sum) getFactorMap() map[string]int {
-	return getFactorMapForFunc[Evaluable[int], Evaluable[int]](s.PairForm, sum)
+func (s *Sum) getFactorMap() map[string]Numeric {
+	return getFactorMapForFunc[Evaluable[Numeric], Evaluable[Numeric]](s.PairForm, sum)
 }
 
-func (s *Sum) Evaluate() int {
+func (s *Sum) Evaluate() Numeric {
 	return s.first.Evaluate() + s.second.Evaluate()
 }
 
-func getFactorMapForFunc[T, U Evaluable[int]](pf *PairForm[T, U], op func(int, int) int) map[string]int {
-	factorMap := make(map[string]int)
+func getFactorMapForFunc[T, U Evaluable[Numeric]](pf *PairForm[T, U], op func(Numeric, Numeric) Numeric) map[string]Numeric {
+	factorMap := make(map[string]Numeric)
 	firstChildMap := pf.GetFirst().getFactorMap()
 	secondChildMap := pf.GetSecond().getFactorMap()
 
@@ -146,10 +146,10 @@ func getFactorMapForFunc[T, U Evaluable[int]](pf *PairForm[T, U], op func(int, i
 }
 
 type Factor struct {
-	*PairForm[Evaluable[int], Evaluable[int]]
+	*PairForm[Evaluable[Numeric], Evaluable[Numeric]]
 }
 
-func NewProduct(first, second Evaluable[int]) *Factor {
+func NewProduct(first, second Evaluable[Numeric]) *Factor {
 	if typed, ok := first.(*Constant); ok {
 		return NewFactor(typed, second)
 	} else if typed, ok := second.(*Constant); ok {
@@ -160,25 +160,25 @@ func NewProduct(first, second Evaluable[int]) *Factor {
 	}
 }
 
-func NewFactor(factor *Constant, value Evaluable[int]) *Factor {
-	return &Factor{NewPairForm[Evaluable[int], Evaluable[int]](factor, value, NoOperator)}
+func NewFactor(factor *Constant, value Evaluable[Numeric]) *Factor {
+	return &Factor{NewPairForm[Evaluable[Numeric], Evaluable[Numeric]](factor, value, NoOperator)}
 }
 
 func (f *Factor) Copy() Form {
 	return &Factor{f.PairForm.TrueCopy()}
 }
 
-func (f *Factor) getFactorMap() map[string]int {
-	factorMap := make(map[string]int)
+func (f *Factor) getFactorMap() map[string]Numeric {
+	factorMap := make(map[string]Numeric)
 	childMap := f.GetSecond().getFactorMap()
 
 	for k, v := range childMap {
-		factorMap[k] = v * int(f.first.Evaluate())
+		factorMap[k] = v * f.first.Evaluate()
 	}
 
 	return factorMap
 }
 
-func (f *Factor) Evaluate() int {
+func (f *Factor) Evaluate() Numeric {
 	return f.first.Evaluate() * f.second.Evaluate()
 }
